feat: Quick Sale POS mode with phone/name customer lookup
- Add Quick Sale / B2B mode toggle in New Sales Order form
- Quick Sale chains create→confirm→ship→invoice in one click
- Customer lookup: live search by phone number or name with dropdown
- Walk-in / Cash quick-select button for anonymous retail sales
- Warehouse selector in Quick Sale mode (stock deducted instantly)
- Barcode scan zone, camera scanner, batch qty parity with PO page
- Extend Customer interface with phone, email, creditLimit fields

+++ apps/inventory/src/app/api/stock/reserve/route.ts
@@ -7,7 +7,7 @@ const reserveSchema = z.object({
     productId: z.string(),
     warehouseId: z.string(),
     variantId: z.string().optional(),
-    quantity: z.number().int().positive(),
+    quantity: z.number().positive(), // Float supports kg-based reservations (e.g. 2.5 kg)
   })).min(1),
   reference: z.string().min(1),
   expiresAt: z.string().datetime().optional(),
